+++ plugin-maven/CHANGES.md
@@ -5,9 +5,10 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ## [Unreleased]
 ### Added
 * Support Ktlint 1.0.0 ([#1808](https://github.com/diffplug/spotless/pull/1808)).
-
 ### Fixed
 * Added support for plugins when using Prettier version `3.0.0` and newer. ([#1802](https://github.com/diffplug/spotless/pull/1802))
+### Changes
+* Bump default `flexmark` version to latest `0.64.0` -> `0.64.6`. ([#1801](https://github.com/diffplug/spotless/pull/1801))
 
 ## [2.39.0] - 2023-08-29
 ### Added