+++ plugin-maven/CHANGES.md
@@ -4,14 +4,14 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 ### Added
-* Support Ktlint 1.0.0 ([#1808](https://github.com/diffplug/spotless/pull/1808)).
-
 * Add support for biome. The Rome project [was renamed to Biome](https://biomejs.dev/blog/annoucing-biome/).
   The configuration is still the same, but you should switch to the new `<biome>` tag and adjust
   the version accordingly. ([#1804](https://github.com/diffplug/spotless/issues/1804)).
-
 ### Fixed
 * Added support for plugins when using Prettier version `3.0.0` and newer. ([#1802](https://github.com/diffplug/spotless/pull/1802))
+### Changes
+* Bump default `flexmark` version to latest `0.64.0` -> `0.64.8`. ([#1801](https://github.com/diffplug/spotless/pull/1801))
+* Bump default `ktlint` version to latest `0.50.0` -> `1.0.0`. ([#1808](https://github.com/diffplug/spotless/pull/1808))
 
 ## [2.39.0] - 2023-08-29
 ### Added