+++ plugin-maven/CHANGES.md
@@ -7,6 +7,19 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 * Add support for biome. The Rome project [was renamed to Biome](https://biomejs.dev/blog/annoucing-biome/).
   The configuration is still the same, but you should switch to the new `<biome>` tag and adjust
   the version accordingly. ([#1804](https://github.com/diffplug/spotless/issues/1804)).
+* Support configuration of mirrors for P2 repositories ([#1697](https://github.com/diffplug/spotless/issues/1697)):
+  ```
+  <eclipse>
+    <p2Mirrors>
+      <p2Mirror>
+        <prefix>https://download.eclipse.org/</prefix>
+        <url>https://some.internal.mirror/eclipse</url>
+      </p2Mirror>
+    </p2Mirrors>
+  </eclipse>
+  ```
+  Mirrors are selected by prefix match, for example `https://download.eclipse.org/eclipse/updates/4.26/` will be redirected to `https://some.internal.mirror/eclipse/eclipse/updates/4.26/`.
+  The same configuration exists for `<greclipse>` and `<eclipseCdt>`.
 ### Fixed
 * Added support for plugins when using Prettier version `3.0.0` and newer. ([#1802](https://github.com/diffplug/spotless/pull/1802))
 ### Changes